+++ src/DataFactory/DataFactoryV2/DataFlows/RemoveAzureDataFactoryDataFlowCommand.cs
@@ -20,7 +20,7 @@ using Microsoft.Azure.Commands.DataFactoryV2.Properties;
 
 namespace Microsoft.Azure.Commands.DataFactoryV2
 {
-    [Cmdlet("Remove", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "DataFactoryV2DataFlow", DefaultParameterSetName = ParameterSetNames.ByFactoryName, SupportsShouldProcess = true)]
+    [Cmdlet("Remove", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "DataFactoryV2DataFlow", DefaultParameterSetName = ParameterSetNames.ByFactoryName, SupportsShouldProcess = false)]
     [OutputType(typeof(void))]
     [OutputType(typeof(bool))]
     public class RemoveAzureDataFactoryDataFlowCommand : DataFactoryContextActionBaseCmdlet